feat: improve SettingsView layout and add comprehensive tests
- Increase SettingsView frame height from 450px to 520px to prevent scrolling
- Remove deprecated "Load Model" button from UI
- Add memory usage display to model status section
- Streamline model status text using centralized WhisperKit.modelState

Test improvements:
- Add comprehensive unit tests for SettingsView frame dimensions
- Add UI tests for model selection and status verification
- Implement helper functions for reliable test state management
- Handle dynamic UI states during model loading (picker replacement with spinner)
- Test model switching between Tiny and Base models
- Verify proper accessibility identifiers and UI element behavior

The changes ensure SettingsView content is always visible without scrolling
while maintaining robust test coverage for frame constraints and UI functionality.
